fix(usage): delete cost-only rows on downgrade instead of zero-filling
A downgrade that zero-fills NULL token_budget leaves cost-only limits as
token_budget=0 rows, which older enforcement reads as 'block at 0 tokens' —
rejecting every request after a rollback. Delete them instead (they can't
function once cost_budget_cents is dropped).

+++ backend/onyx/db/models.py
@@ -5467,9 +5467,11 @@ class UserUsage(Base):
         String, nullable=False, default="", server_default=""
     )
 
-    input_tokens: Mapped[int] = mapped_column(Integer, nullable=False)
-    output_tokens: Mapped[int] = mapped_column(Integer, nullable=False)
-    cache_read_tokens: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
+    input_tokens: Mapped[int] = mapped_column(BigInteger, nullable=False)
+    output_tokens: Mapped[int] = mapped_column(BigInteger, nullable=False)
+    cache_read_tokens: Mapped[int] = mapped_column(
+        BigInteger, nullable=False, default=0
+    )
     cost_cents: Mapped[float] = mapped_column(Float, nullable=False, default=0.0)
 
     created_at: Mapped[datetime.datetime] = mapped_column(
